+++ pages/api/create-project.ts
@@ -4,7 +4,7 @@ import { NextRequest, NextResponse } from 'next/server'
 import uuid from 'react-uuid'
 import { createAdminClient, createEdgeClient } from './_db'
 import { projectSlugify, toTitleCase } from '@/utils/formatting'
-import { ProjectParams } from '@/app/create/create-project-form'
+import { ProjectParams } from '@/utils/upsert-project'
 import { getPrizeCause, updateProjectCauses } from '@/db/cause'
 import { getProposalValuation, getMinIncludingAmm } from '@/utils/math'
 import { insertBid } from '@/db/bid'
