+++ utils/upsert-project.ts
@@ -1,5 +1,23 @@
-import { Project, TOTAL_SHARES } from '@/db/project'
+import { Cause, MiniCause } from '@/db/cause'
+import { TOTAL_SHARES } from '@/db/project'
 import { SupabaseClient } from '@supabase/supabase-js'
+import { JSONContent } from '@tiptap/core'
+import uuid from 'react-uuid'
+
+export type ProjectParams = {
+  title: string
+  subtitle: string | null
+  minFunding?: number
+  fundingGoal?: number
+  verdictDate: string
+  description?: JSONContent | string
+  location: string
+  selectedCauses: MiniCause[]
+  selectedPrize: Cause | null
+  founderPercent: number
+  agreedToTerms: boolean
+  lobbying: boolean
+}
 
 export async function giveCreatorShares(
   supabase: SupabaseClient,
@@ -37,9 +55,30 @@ export async function upvoteOwnProject(
   }
 }
 
-export async function saveProject(supabase: SupabaseClient, project: Project) {
-  const { error } = await supabase.from('projects').insert([project])
-  if (error) {
-    console.error(error)
+export function createUpdateFromParams(projectParams: ProjectParams) {
+  const {
+    title,
+    subtitle,
+    minFunding,
+    fundingGoal,
+    verdictDate,
+    description,
+    location,
+    selectedPrize,
+    founderPercent,
+    lobbying,
+  } = projectParams
+  return {
+    title,
+    blurb: subtitle,
+    description,
+    min_funding: minFunding ?? 0,
+    funding_goal: fundingGoal ?? minFunding ?? 0,
+    founder_shares: !!selectedPrize
+      ? (founderPercent / 100) * TOTAL_SHARES
+      : TOTAL_SHARES,
+    auction_close: verdictDate,
+    location_description: location,
+    lobbying,
   }
 }
